launchbot/requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-pinecone==0.2.11
langchain-tests==0.3.20
langchain-text-splitters==0.3.9
langsmith==0.4.10
openai==1.98.0
pinecone==7.3.0
beautifulsoup4==4.13.4
requests==2.32.4
tiktoken==0.9.0
tqdm==4.67.1
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
